% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxr.R
\name{\%OR\%}
\alias{\%OR\%}
\title{Bitwise OR - a simple wrapper around \code{\link{bitwOr}} that is
a little nicer to look at.}
\usage{
a \%OR\% b
}
\arguments{
\item{a,b}{scalar integers}
}
\description{
Bitwise OR - a simple wrapper around \code{\link{bitwOr}} that is
a little nicer to look at.
}
\examples{
tb$BLUE
tb$BLUE \%OR\% tb$BOLD
bitwOr(tb$BLUE, tb$BOLD)
}

