% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_add_description.R
\name{box_add_description}
\alias{box_add_description}
\title{Add description to Box file}
\usage{
box_add_description(file_id, description)
}
\arguments{
\item{file_id}{\code{numeric} or \code{character}, file ID at Box}

\item{description}{\code{character}, description caption for the file}
}
\value{
S3 object with class \link[=boxr_S3_classes]{boxr_file_reference}.
}
\description{
Files hosted at Box can have small descriptions.
These can be useful to explain the contents of files, or even to
leave 'git commit' style messages about the latest changes made to them.
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
