% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_collab.R
\name{box_collab_create_internal}
\alias{box_collab_create_internal}
\title{Collaboration creation station}
\usage{
box_collab_create_internal(item, accessible_by, role, can_view_path = FALSE)
}
\description{
Collaboration creation station
}
\keyword{internal}
