% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_collab.R
\name{box_collab_delete}
\alias{box_collab_delete}
\title{Delete Box collaboration}
\usage{
box_collab_delete(collab_id)
}
\arguments{
\item{collab_id}{\code{character} ID for Box collaboration}
}
\value{
Invisible \code{NULL}, called for side-effects.
}
\description{
Delete Box collaboration
}
