% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_collab.R
\name{box_collab_get}
\alias{box_collab_get}
\title{Get Box collaborations}
\usage{
box_collab_get(dir_id = NULL, file_id = NULL)
}
\arguments{
\item{dir_id}{\code{numeric} or \code{character}, folder ID at Box.}

\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}
}
\value{
Object with S3 class
\code{\link[=boxr_S3_classes]{boxr_collab_list}}.
}
\description{
Retrieve information on all collaborations on a file or folder.
}
\details{
You must specify exactly one of \code{dir_id} or \code{file_id}.

This returns an object with S3 class \code{\link[=boxr_S3_classes]{boxr_collab_list}};
this is a list containing the response from the API. You can use
\code{as_tibble()} or \code{as.data.frame()} on this return-object to convert to
a tibble or data frame.
}
