% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_comment.R
\name{box_comment_create}
\alias{box_comment_create}
\alias{box_comment_get}
\title{Create/delete Box comments}
\usage{
box_comment_create(file_id = NULL, msg, comment_id = NULL)

box_comment_get(file_id)
}
\arguments{
\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}

\item{msg}{\code{character} the comment Note: tagging people with the
\@user pattern is \emph{not} supported at this time.}

\item{comment_id}{\code{numeric} or \code{character}, comment ID at Box.}
}
\value{
Invisible, the \code{comment_id} of the newly created Box comment.

A \code{data.frame} with a row for each comment.
}
\description{
Comments can be associated with files or other comments. If a \code{comment_id}
is used it will be posted as a reply. Retreiving comments en bulk is only
possible on a \code{file_id}.
}
\examples{
\dontrun{
box_comment(12345, "Rmd report is ready for review")
box_comment(comment_id = 98756, msg = "Responsed to your change requests!")
}
\dontrun{
box_comment_get(12345)
}
}
