% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_collab.R
\name{box_create_collab_internal}
\alias{box_create_collab_internal}
\title{Collaboration creation station}
\usage{
box_create_collab_internal(item, accessible_by, role, can_view_path = FALSE)
}
\description{
Collaboration creation station
}
\keyword{internal}
