% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr_delete_restore.R
\name{box_delete_file}
\alias{box_delete_file}
\alias{box_delete_folder}
\alias{box_restore_file}
\alias{box_restore_folder}
\title{Move files to and from the box.com trash folder}
\usage{
box_delete_file(file_id)

box_delete_folder(dir_id)

box_restore_folder(dir_id)

box_restore_file(file_id)
}
\arguments{
\item{file_id}{The box.com id for the file that you'd like delete/restore}

\item{dir_id}{The box.com id for the folder that you'd like delete/restore}
}
\value{
Nothing. Used for their side effects.
}
\description{
\code{box_delete_file} moves a file to the trash folder.
\code{box_delete_folder} moves a whole folder (and all of it's contents) to
the trash folder.
\code{box_restore_file} moves a file from the trash, to wherever it was
before it was 'deleted'/moved to trash.
\code{box_delete_folder} does the same thing for a folder.
}

