% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_collab.R
\name{box_dir_invite}
\alias{box_dir_invite}
\title{Invite collaboration}
\usage{
box_dir_invite(
  dir_id,
  user_id,
  login = NULL,
  role = "viewer",
  can_view_path = FALSE
)
}
\arguments{
\item{dir_id}{\code{numeric} or \code{character}, folder ID at Box.}

\item{user_id}{\code{character} ID for Box user-account to invite.}

\item{login}{\code{character} email address of account to invite, if specified
will be used instead of \code{user_id}.}

\item{role}{\code{character} role of the collaborator; default is \code{"viewer"}.}

\item{can_view_path}{\code{logical} indicates to allow the collaborator to
navigate parent-folders at Box.}
}
\value{
Invisible \code{list()}.
}
\description{
\code{box_dir_invite()} is deprecated in favor of \code{box_collab_create()}.
}
\keyword{\lifecycle{deprecated}}
\keyword{deprecated}
