% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr.R
\name{box_dl}
\alias{box_dl}
\alias{box_read}
\title{Download a file from box.com}
\usage{
box_dl(file_id, overwrite = FALSE, local_dir = getwd(), filename = NULL)

box_read(file_id)
}
\arguments{
\item{file_id}{The box.com id for the file that you'd like to download}

\item{overwrite}{\code{logical}. Should existing files with the same name be
overwritten?}

\item{local_dir}{The local directory which you'd like \code{dir_id} to
correspond to. If you're not interested in mapping local to hosted
directories, this isn't terribly important.}

\item{filename}{The filename for the local version of the file. The default,
\code{NULL}, uses the name from box.com.}
}
\value{
\code{TRUE}. Used for it's side-effect (a downloaded file)
}
\description{
Takes the \code{id} of a file hosted on box.com, downloads it, either
writing it to disk (\code{box_dl}), or loading it as an \code{R} object
(\code{box_read}).
}

