% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr.R
\name{box_fetch}
\alias{box_fetch}
\title{Synchronize the directory structure and file contents of a box.com folder}
\usage{
box_fetch(dir_id, local_dir = getwd(), recursive = TRUE, overwrite = TRUE)
}
\arguments{
\item{dir_id}{The box.com id for the folder that you'd like to query}

\item{local_dir}{The local directory which you'd like \code{dir_id} to
correspond to. If you're not interested in mapping local to hosted
directories, this isn't terribly important.}

\item{recursive}{\code{logical}. Should the call include subdirectories and
thier contents?}

\item{overwrite}{\code{logical}. Should existing files with the same name be
overwritten?}
}
\value{
Nothing. Used for its side-effects.
}
\description{
Download the contents of a box.com directory
}
\details{
The box.com API does not have direct support for downloading more
than one file. With \code{recursive} set to \code{false}, \code{box_pull}
will download the files, but not subdirectories of the folder specified by
\code{dir_id}. If \code{recursive == TRUE}, then it will download every file
and folder in the directory tree. Because R has to make recursive API calls
to explore the directory structure, and then iterate through each file it
finds, this option can be rather slow.
}

