% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_auth.R
\name{box_fresh_auth}
\alias{box_fresh_auth}
\title{Reauthorise a Problematic box.com connection}
\usage{
box_fresh_auth(cache = "~/.boxr-oauth", ...)
}
\arguments{
\item{cache}{Passed to \code{cache} in \code{\link{httr}}.}

\item{...}{Passed to \code{\link{box_auth}}}
}
\description{
Very simply, deletes the old token file before trying to re-authorise. This 
is often the solution to authorisation problems raised by users!
}
\seealso{
\code{\link{box_auth}} for the usual method of authorisation, and
  \code{\link{box_auth_on_attach}} for a lazy one.
}

