% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_collab.R
\name{box_get_collab}
\alias{box_get_collab}
\title{Get existing collaborations}
\usage{
box_get_collab(dir_id = NULL, file_id = NULL)
}
\arguments{
\item{dir_id}{\code{numeric} or \code{character}, folder ID at Box.}

\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}
}
\value{
Invisible \code{data.frame} with one row per collaboration.
}
\description{
You must specify either \code{dir_id} or \code{file_id}, if both a specified \code{dir_id} is used.
}
