% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_collab.R
\name{box_get_collab_dir}
\alias{box_get_collab_dir}
\alias{box_get_collab_file}
\title{Get the existing collaborations on a file or folder.}
\usage{
box_get_collab_dir(dir_id)

box_get_collab_file(file_id)
}
\arguments{
\item{dir_id}{\code{numeric} or \code{character}, folder ID at Box.}

\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}
}
\description{
Get the existing collaborations on a file or folder.
}
