% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{box_ls}
\alias{box_ls}
\title{List files in a Box directory}
\usage{
box_ls(dir_id = box_getwd(), limit = 100, max = Inf, fields = NULL)
}
\arguments{
\item{dir_id}{\code{character} or \code{double}, folder id at \href{https://box.com}{box.com}}

\item{limit}{\code{integer}, maximum number of entries to retrieve per query-page}

\item{max}{\code{integer}, maximum number of entries to retrieve in total}

\item{fields}{\code{character}, fields to return; the default
value, \code{NULL`` will return all possible columns: }modified_at\code{, }content_modified_at\code{, }name\code{, }id\code{, }type\code{, }sha1\code{ ,}size\code{, }owned_by\code{, }path_collection\code{, }description`}
}
\value{
S3 object wuth class \code{boxr_object_list} describing the contents
of the folder specified by \code{dir_id}.
}
\description{
Non-recursive
}
\seealso{
\code{\link[=box_fetch]{box_fetch()}} and \code{\link[=box_push]{box_push()}} for synchronizing the contents of
local and remote directories, \code{\link[=list.files]{list.files()}} for examining the contents of
local directories.
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}, Ian Lyttle
\email{ian.lyttle@schneider-electric.com}, and Alec Wong \email{aw685@cornell.edu}
}
