% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{box_ls_short}
\alias{box_ls_short}
\title{Shorter version of box_ls.}
\usage{
box_ls_short(dir_id = box_getwd(), limit = 100, max = Inf)
}
\arguments{
\item{dir_id}{The box.com id for the folder that you'd like to query}

\item{limit}{Maximum number of entries to retrieve per query-page}

\item{max}{Maximum number of entries to retrieve in total}
}
\value{
A data.frame describing the contents of the the folder specified by 
  \code{dir_id}. Non recursive.
}
\description{
Shorter version of box_ls.
}
\seealso{
\code{\link{box_fetch}} and \code{\link{box_push}} for synchronizing
  the contents of local and remote directories. \code{\link{list.files}} for
  examining the contents of local directories.
}
\author{
Alec Wong \email{aw685@cornell.edu}
}
