% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{box_open_file}
\alias{box_open_file}
\alias{box_open_dir}
\title{Open a browser to view a file or folder on Box's web interface}
\usage{
box_open_file(file_id)

box_open_dir(dir_id)
}
\arguments{
\item{file_id}{\code{character} a Box file ID}

\item{dir_id}{\code{character} a Box folder ID}
}
\description{
Thin wrapper of \code{browseURL} to make bouncing between R and Box a breeze.
}
\examples{
\dontrun{
box_open_file(12345)

box_open_folder(0)
}
}
