% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_search.R
\name{box_search_pagination}
\alias{box_search_pagination}
\title{This 'pagination'. The box.com API returns a maximum of 200 results; this
will keep making requests until they're all in}
\usage{
box_search_pagination(url, max = 200)
}
\description{
This 'pagination'. The box.com API returns a maximum of 200 results; this
will keep making requests until they're all in
}
\keyword{internal}

