% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr.R
\name{box_setwd}
\alias{box_getwd}
\alias{box_setwd}
\title{set directory}
\usage{
box_setwd(dir_id)

box_getwd()
}
\arguments{
\item{dir_id}{The box.com id for the folder that you'd like to query}
}
\value{
Nothing. Used for its side-effects.
}
\description{
set directory
}

