% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_source.R
\name{box_source}
\alias{box_source}
\title{Source R code from a Box file}
\usage{
box_source(file_id)
}
\arguments{
\item{file_id}{The box.com id for the file that you'd like to download}
}
\description{
\code{box_source} will download a remote R script stored on box.com, and then
attempt to execute it, using \code{\link[=source]{source()}}.
}
\seealso{
\itemize{
\item \code{\link[=box_dl]{box_dl()}} for saving files to disk,
\item \code{\link[=box_save]{box_save()}} for working with R workspaces, and
\item \code{\link[=box_read]{box_read()}} for reading files into memory as R objects.
}
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
