% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr__internal_misc.R
\name{box_terminal_http_codes}
\alias{box_terminal_http_codes}
\title{Return common Box API client errors for terminating httr::RETRY()}
\usage{
box_terminal_http_codes()
}
\value{
Numeric with HTTP status codes.
}
\description{
This function returns a subset of known Box API error codes, based on the
\href{https://developer.box.com/guides/api-calls/permissions-and-errors/common-errors/}{Box API docs}.
This function is only intended to be used as an argument to httr::RETRY to prevent
successive API requests when the orignal request succeeded but returned a error unrelated
to establishing a connection.
}
