% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr__internal_update_upload.R
\name{box_upload_new}
\alias{box_upload_new}
\alias{box_update_file}
\alias{box_update_file}
\title{Upload or update a file}
\usage{
box_upload_new(dir_id, file, pb = FALSE)

box_update_file(file_id, file, dir_id, pb = FALSE)
}
\arguments{
\item{dir_id}{The box.com id for the folder that you'd like to upload to}

\item{file}{A path to a file stored locally}

\item{file_id}{the box.com id of the file you'd like to update}
}
\value{
The \code{\link{httr}} object returned by the api call
}
\description{
Internal functions used by \code{\link{box_ul}} (who's use is recommended as
an alternative to these functions).
}
\details{
The box.com api requires different API calls to upload a new file, and to
upload a new version of a file which already exists (incrementing the version
number).

\code{box_upload_new} make the API call to upload a new file.
\code{box_update_file} makes the API call to update an existing file.
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
\keyword{internal}
