% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_auth.R
\name{box_user_id}
\alias{box_user_id}
\title{Get Box user-id}
\usage{
box_user_id()
}
\value{
\code{character} value of Box user-id
}
\description{
Get Box user-id
}
\examples{
# throws error if environment-variable not set
\dontrun{
  box_user_id()
}
}
