% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_file_versions.R
\name{box_version_api}
\alias{box_version_api}
\title{Access Box version API}
\usage{
box_version_api(file_id)
}
\arguments{
\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}
}
\value{
Object with S3 class \code{"boxr_version_list"}
}
\description{
Use this function to access the response-content for the
versions API endpoint.
}
\keyword{internal}
