% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr__internal_get.R
\name{boxGet}
\alias{boxGet}
\title{Issue a get request for a file stored on box.com}
\usage{
boxGet(file_id, local_file, version_id = NULL, version_no = NULL,
  download = FALSE, pb = FALSE)
}
\description{
This internal function is shared by \code{\link{box_dl}}, and the 
\code{\link{box_read}} family of functions, to issue GET requests, while
handling things like version numbers etc. It can be used to download a file,
or just read it into memory.
}
\keyword{internal}

