% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxr.R
\name{boxr_change_cell}
\alias{boxr_change_cell}
\title{Set many cells}
\usage{
boxr_change_cell(x, y, ch, fg, bg)
}
\arguments{
\item{x,y}{x and y position of the cells so c(x[i], y[i]) is the ith cell}

\item{ch}{A single or vector of character codes (integer)}

\item{fg,bg}{foregreound and background colours (see \code{\link{tb}}
for a set of values)}
}
\description{
Vectorised version of \code{\link{tb_change_cell}}; any elements
in \code{x}, \code{y} \code{ch} \code{fg} and \code{bg} may be
vectors with length greater than 1, but only one length greater
than 1 is allowed, and \code{x} and/or \code{y} must have length
greater than 1 if anything else does.
}

