% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxr.R
\name{boxr_change_rect}
\alias{boxr_change_rect}
\title{Set a rectangle of cells}
\usage{
boxr_change_rect(x, y, ch, fg, bg)
}
\arguments{
\item{x}{a pair of points being the x edge of the rectangle}

\item{y}{a pair of points being the y edge of the rectangle}

\item{ch}{A single integer character code}

\item{fg,bg}{Single foreground and background colours (see \code{\link{tb}}
for a set of values)}
}
\description{
Wrapper around \code{boxr_change_cell} to set a rectangle to a
single colour and character (it is possible to pass vectors here
in but not recommended as the interface may change).
}

