% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{boxr_options}
\alias{boxr_options}
\title{Get boxr options}
\usage{
boxr_options()
}
\value{
\code{list}, current values of boxr options, with elements:

\describe{
\item{\code{boxr.interactive}}{\code{logical}, indicates if boxr is running in interactive mode.}
\item{\code{boxr.progress}}{\code{logical}, indicates to use progress-bars, if available.}
\item{\code{boxr.verbose}}{\code{logical}, indicates if boxr will use \code{\link[=cat]{cat()}} to print to the console.
Setting to \code{TRUE} may cause problems with \code{knitr}.}
\item{\code{boxr.wd}}{\code{list}, containing information on the Box working-directory:
\code{id} \code{(numeric)}, and \code{name} \code{(character)}.}
\item{\code{boxr.wd.path}}{\code{character}, path to the Box working-directory.}
\item{\code{boxr.token}}{Object with S3 class \code{Token2.0} (\code{\link[httr:Token2.0]{httr::Token2.0}}).}
}
}
\description{
This function gets the values of boxr's global options.
}
\details{
Options can be set in the usual way, using \code{\link[=options]{options()}}.
}
