% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr__internal_dir_comparison.R
\name{create_loc_dir_df}
\alias{create_loc_dir_df}
\title{Create a data.frame of metadata of the contents of a local directory}
\usage{
create_loc_dir_df(local_dir = getwd())
}
\arguments{
\item{local_dir}{The local directory which you'd like \code{dir_id} to
correspond to. If you're not interested in mapping local to hosted
directories, this isn't terribly important.}
}
\value{
A data.frame of metadata.
}
\description{
Create a data.frame of metadata of the contents of a local directory
}
\keyword{internal}
