% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr.R
\name{downloadDirFiles}
\alias{downloadDirFiles}
\alias{uploadDirFiles}
\title{Single Directory Operations for Downloading and Uploading mutiple Files}
\usage{
downloadDirFiles(dir_id, local_dir = getwd(), overwrite = TRUE)

uploadDirFiles(dir_id, local_dir = getwd())
}
\arguments{
\item{dir_id}{The box.com id for the folder that you'd like to query}

\item{local_dir}{The local directory which you'd like \code{dir_id} to
correspond to. If you're not interested in mapping local to hosted
directories, this isn't terribly important.}
}
\value{
\code{TRUE} for a successful sync, \code{NULL} if the box.com folder
is empty.
}
\description{
Download or upload the contents of a box.com directory, not including
subdirectories
}
\keyword{internal}

