% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_add_description.R
\name{box_add_description}
\alias{box_add_description}
\title{Add description to a Box file}
\usage{
box_add_description(file_id, description)
}
\arguments{
\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}

\item{description}{\code{character}, description caption for the file.}
}
\value{
Object with S3 class \code{\link[=boxr_S3_classes]{boxr_file_reference}}.
}
\description{
These functions will attach a description or comment to a Box file. A new
description will overwrite an existing one.
}
\details{
Files hosted at Box can have small text-descriptions that
you can be use to annotate files, or even to
}
