% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_add_description.R
\name{box_add_description}
\alias{box_add_description}
\title{Add a description to a file hosted remotely on box.com}
\usage{
box_add_description(file_id, description)
}
\arguments{
\item{file_id}{The box.com id for the file that you'd like to download}

\item{description}{The description which you'd like to add to the file. 
\code{\link{character}}.}
}
\value{
An object of class 
  \code{\link[=boxr_S3_classes]{boxr_file_reference}}.
}
\description{
Files hosted on box.com can have small descriptions underneath their
filenames. This can be useful to explain the contents of the file, or even to
leave 'git commit' style messages about the latest changes made to them.
}

