% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_add_description.R
\name{box_add_description}
\alias{box_add_description}
\title{Add description to Box file}
\usage{
box_add_description(file_id, description)
}
\arguments{
\item{file_id}{\code{numeric} or \code{character}, file ID at Box}

\item{description}{\code{character}, description caption for the file}
}
\value{
S3 object with class \code{\link[=boxr_S3_classes]{boxr_folder_reference}}.
}
\description{
This function will attach a description to a Box file; it will
overwrite the Box file's existing description.
}
\details{
Files hosted at Box can have small text-descriptions that
you can be use to annotate files, or even to
leave 'git commit' style messages.
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
