% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr.R
\name{box_auth}
\alias{box_auth}
\title{Authenticate box.com account}
\usage{
box_auth(client_id = "", client_sc = "", interactive = TRUE,
  use_oob = getOption("httr_oob_default"), as_header = TRUE,
  cache = "~/.boxr-oath", write.Renv = TRUE, reset.Renv = FALSE)
}
\arguments{
\item{client_id}{The client id for the account you'd like to use.
\code{character}.}

\item{client_sc}{The client secret for the account you'd like to use.
\code{character}.}

\item{interactive}{\code{logical}. Should the authorization process happen
interactively (requiring user input to the R console, and/or a visit to
box.com)?}

\item{use_oob}{Passed to \code{oob} in \code{\link{httr}}.}

\item{as_header}{Passed to \code{as_header} in \code{\link{httr}}.}

\item{cache}{Passed to \code{cache} in \code{\link{httr}}.}

\item{write.Renv}{\code{logical}. If they were missing, and an OAuth2.0 token
was obtained, should \code{client_id} and \code{client_sc} be written to
\code{.Renvirons} in  your \code{HOME} directory? (Note: The \code{HOME} dir
is not neccesarily that returned by \code{geetwd()}.)}

\item{reset.Renv}{\code{logical}. Should existing values for \code{client_id}
and \code{client_sc} in \code{.Renvirons} be ignored?}
}
\value{
Invoked for it's side effect; OAuth2.0 connection to the box.com
API.
}
\description{
Authenticate box.com account
}

