% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_auth.R
\name{box_auth_jwt}
\alias{box_auth_jwt}
\title{Authenicate a box.com account with a JWT (JSON Web Token)}
\usage{
box_auth_jwt(config_file = NULL, user_id = NULL)
}
\arguments{
\item{config}{Path to JSON config file.}

\item{user}{User ID number for Box account authorization.}
}
\description{
More doc-ing needed on setting up a JWT app, very similar to O-Auth2.
}
