% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_auth.R
\name{box_auth_jwt}
\alias{box_auth_jwt}
\title{Authenicate a box.com account with a JWT (JSON Web Token)}
\usage{
box_auth_jwt(user_id = NULL, config_file = NULL)
}
\arguments{
\item{user_id}{\code{character}, the user ID for the account to use.}

\item{config_file}{\code{character} path to JSON config file.}
}
\description{
Alternative option for accessing the Box API. Useful on servers.
}
