% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_auth.R
\name{box_auth_service}
\alias{box_auth_service}
\title{Authenticate to Box (service)}
\usage{
box_auth_service(token_file = NULL, token_text = NULL,
  account_id = NULL)
}
\arguments{
\item{token_file}{\code{character}, path to JSON token-file}

\item{token_text}{\code{character}, JSON text. If this is provided,
\code{token_file} is ignored.}

\item{account_id}{\code{character}, the ID for the account to use.}
}
\description{
Alternative option for accessing the Box API. Useful on servers.
}
