% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_auth.R
\name{box_auth_service}
\alias{box_auth_service}
\title{Authenticate to Box (service-app)}
\usage{
box_auth_service(token_file = NULL, token_text = NULL,
  account_id = NULL)
}
\arguments{
\item{token_file}{\code{character}, path to JSON token-file.}

\item{token_text}{\code{character}, JSON text. If this is provided,
\code{token_file} is ignored.}

\item{account_id}{\code{character}, the ID for the account to use; default is
to use the service-account for the app.
Generally, this argument is not specified; the default behavior is used.}
}
\value{
Invisible \code{NULL}, called for side-effects.
}
\description{
How you authenticate to Box depends on how the Box-app through which you
connect. A Box service-app can be useful for unattended jobs that need
access to only a limited part of Box, e.g. one folder.

Use this function to access Box using a service-app.

To access a service-app, you will need a JSON web-token (JWT),
generated by your Box-admin team. If you have a personal Box account, \emph{you}
are your Box-admin team. You specify the JWT either as \code{token_file},
the path to the JWT file, or as \code{token_text}, the text of the JWT.

Using JWT-authenitcation is more convienient than using standard OAuth2
authentication, as you do not have to go through the "OAuth Dance". This
convenience brings additional considerations because the JWT file gives
its bearer uninhibited access to anything the Box service-app can access:
\itemize{
\item give the service-account access to as little information as you need it
to have, e.g. a single folder.
\item keep the JWT file secure.
}
}
\details{
The default behavior of a service-app is to act on behalf of the
service-account associated with the service-app. This is different
from an interactive-app, which acts on behalf of the Box user who
authenticates to it.

To use a service-app on a folder belonging to a Box user, either
the Box user has to invite the service-account to collaborate on a
folder belonging to the user, or the service-account has to invite the
Box user to collaborate on a folder belonging to the service-account.

In either case, you can use \code{box_dir_invite()}.

For more details on Box service-apps, including how to create them, and
service-app-based workflows, please read \code{vignette("boxr-app-service")}.
}
\section{Side-effects}{


This function has some side effects:
\itemize{
\item some global \code{\link[=options]{options()}} are set for your session to manage the token.
\item a message is printed to the console.
}
}

\seealso{
\code{\link[=box_auth_]{box_auth_()}} for authenticating to interactive-apps,
\code{\link[=box_dir_invite]{box_dir_invite()}} for inviting a different account to collaborate on
a Box folder.
}
