% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{box_browse}
\alias{box_browse}
\title{Open a Box directory or file in browser}
\usage{
box_browse(dir_id = NULL, file_id = NULL)
}
\arguments{
\item{dir_id}{\code{numeric} or \code{character}, folder ID at Box.}

\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}
}
\value{
Invisible \code{NULL}, called for side-effects.
}
\description{
Thin wrapper of \code{utils::browseURL()} to make bouncing between R and Box a breeze.
}
\examples{
\dontrun{
  box_browse(0) # root folder on Box
  box_browse(file_id = 12345)
}

}
