% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_collab.R
\name{box_delete_collab}
\alias{box_delete_collab}
\title{Delete a collaboration on Box.}
\usage{
box_delete_collab(collab_id)
}
\arguments{
\item{collab_id}{\code{numeric} ID for Box collaboration}
}
\value{
the Box API response object
}
\description{
Delete a collaboration on Box.
}
