% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{box_dir_create}
\alias{box_dir_create}
\title{Create a Box directory}
\usage{
box_dir_create(dir_name, parent_dir_id = box_getwd())
}
\arguments{
\item{dir_name}{\code{character}, name for new folder at Box.}

\item{parent_dir_id}{\code{character} or \code{numeric},
ID for the parent folder at Box.}
}
\value{
S3 object with class \code{\link[=boxr_S3_classes]{boxr_folder_reference}}.
}
\description{
This will create a new folder at Box, with name \code{dir_name},
in the Box folder with ID \code{parent_dir_id}.
}
\seealso{
\code{\link[=box_delete_folder]{box_delete_folder()}} to move Box folders to trash,
\code{\link[=box_ls]{box_ls()}} to list files in a Box folder.
}
