% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_invite.R
\name{box_dir_invite}
\alias{box_dir_invite}
\title{Invite a collaborator to a Box folder}
\usage{
box_dir_invite(dir_id, account_id = box_user_id(), login = NULL,
  role = "viewer", can_view_path = FALSE)
}
\arguments{
\item{dir_id}{\code{numeric} or \code{character}, folder ID at Box.}

\item{account_id}{\code{character} ID for Box account to invite}

\item{login}{\code{character} email address of account to invite, can be used instead of
\code{account_id}}

\item{role}{\code{character} role of the collaborator}

\item{can_view_path}{\code{logical} indicates to allow the collaborator to navigate
parent-folders at Box.}
}
\value{
Invisible \code{list()} containing collaboration-information.
}
\description{
default to lowest permissions
}
