% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr_upload_download.R
\name{box_dl}
\alias{box_dl}
\alias{box_read}
\alias{box_ul}
\title{Download and upload individual files from box.com}
\usage{
box_dl(file_id, local_dir = getwd(), overwrite = FALSE, filename = NULL)

box_ul(dir_id = getOption("boxr.wd")$id, file)

box_read(file_id)
}
\arguments{
\item{file_id}{The box.com id for the file that you'd like to download}

\item{local_dir}{A file path to a local directory which you'd like the file
to be downloaded to.}

\item{overwrite}{\code{logical}. Should existing files with the same name be
overwritten?}

\item{filename}{Optional. An alternate filename for the local version of the
file. The default, \code{NULL}, uses the name from box.com.}

\item{dir_id}{If uploading, the box.com folder id that you'd like to upload
to.}

\item{file}{the path to the local file that you'd like to upload (if there is
one)}
}
\value{
\code{TRUE}. Used for it's side-effect (a downloaded file)
}
\description{
Functions to download(\code{box_dl}), and upload (J\code{box_ul}), files, and
read remote files straight into memory as R objects (\code{box_read}).
}
\details{
\code{box_dl} takes the \code{id} of a file hosted on box.com, downloads
it and writes it to disk.

\code{box_read} does the same, but reads it into memory as an \code{R}
object. This can be useful, for example, to read a \code{.csv} file into
memory as a \code{\link{data.frame}}.

\code{box_ul} uploads a file stored locally to a specified box.com folder.
If a file with the same name already exists, it will store a new version of
the file.
}

