% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_auth.R
\name{box_fresh_auth}
\alias{box_fresh_auth}
\title{Re-authenticate to Box (interactive-app)}
\usage{
box_fresh_auth(cache = "~/.boxr-oauth", ...)
}
\arguments{
\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}, \code{FALSE} means
don't cache, and \code{NA} means to guess using some sensible heuristics.
A string means use the specified path as the cache file.}

\item{...}{Other arguments passed to \code{\link[=box_auth]{box_auth()}}.}
}
\description{
Deletes the cached token-file before trying to re-authenticate. This
is often the solution to authentication problems.
}
\seealso{
\code{\link[=box_auth]{box_auth()}} for the usual method of authorisation
}
