% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{box_ls}
\alias{box_ls}
\title{Obtain a data.frame describing the contents of a directory}
\usage{
box_ls(dir_id)
}
\arguments{
\item{dir_id}{The box.com id for the folder that you'd like to query}
}
\value{
A data.frame describing the contents of the the folder specified by
\code{dir_id}. Non recursive.
}
\description{
Obtain a data.frame describing the contents of a directory
}

