% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{box_ls}
\alias{box_ls}
\title{List files in a Box directory}
\usage{
box_ls(dir_id = box_getwd(), limit = 100, max = Inf, fields = NULL)
}
\arguments{
\item{dir_id}{The box.com id for the folder that you'd like to query}

\item{limit}{Maximum number of entries to retrieve per query-page}

\item{max}{Maximum number of entries to retrieve in total}

\item{fields}{Specify what fields to query as a character vector. The default
value NULL will return all possible columns: \code{modified_at},
\code{content_modified_at}, \code{name}, \code{id}, \code{type}, \code{sha1} ,\code{size},
\code{owned_by}, \code{path_collection}, \code{description}}
}
\value{
A data.frame describing the contents of the the folder specified by
\code{dir_id}. Non recursive.
}
\description{
List files in a Box directory
}
\seealso{
\code{\link[=box_fetch]{box_fetch()}} and \code{\link[=box_push]{box_push()}} for synchronizing the contents of
local and remote directories. \code{\link[=list.files]{list.files()}} for examining the contents of
local directories.
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}, Ian Lyttle
\email{ian.lyttle@schneider-electric.com}, and Alec Wong \email{aw685@cornell.edu}
}
