% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_file_versions.R
\name{box_previous_versions}
\alias{box_previous_versions}
\alias{box_version}
\title{Get details about versions of a Box file}
\usage{
box_previous_versions(file_id)

box_version(file_id)
}
\arguments{
\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}
}
\value{
\describe{
\item{\code{box_previous_versions()}}{
\code{data.frame} describing all versions of file}
\item{\code{box_version()}}{
\code{integer} version number of most-recent version of file}
}
}
\description{
Box uses file versioning, but the API does not explicitly provide version
numbers. These functions use \code{modified_date} as a proxy to determine a
version number (\code{version_no}), which you can use with \code{\link[=box_dl]{box_dl()}} and
\code{\link[=box_read]{box_read()}}.
}
\details{
\code{box_previous_versions()} gets information on all versions of a file.

\code{box_version()} gets the version number of the most-recent version.
}
\references{
This function is a light wrapper of the
\href{https://developer.box.com/docs}{box.com} API \code{versions} method.

\url{https://developers.box.com/docs/#files-view-versions-of-a-file}
}
\seealso{
\code{\link[=box_dl]{box_dl()}}, \code{\link[=box_read]{box_read()}}
}
