% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_file_versions.R
\name{box_previous_versions}
\alias{box_previous_versions}
\alias{box_current_version}
\title{Get details about versions of a Box file}
\usage{
box_previous_versions(file_id)

box_current_version(file_id)
}
\arguments{
\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}
}
\value{
\code{data.frame} containing information about previous
versions of the file (if available).
}
\description{
Box uses file versioning, but the API does not explicitly provide version numbers.
These functions use \code{modified_date} as a proxy and allow programatic version access
in \code{box_dl()} and \code{box_read()}
}
\details{
\itemize{
\item \code{box_previous_version()} returns a
\code{data.frame} containing information on a file's previous
versions on Box, the column \code{file_version_id} can be passed to \code{version_id}
in \code{box_dl()} and \code{box_read()}.
If the version of a file is one, then NULL is returned invisibly
along with a helpful message.
\item \code{box_current_version()} returns a \code{integer}, starting from 1, which can be passed
to \verb{version_no = } in \code{box_dl()} and \code{box_read()}.
}
}
\references{
This function is a light wrapper of the
\href{https://developer.box.com/docs}{box.com} API \code{versions} method.

\url{https://developers.box.com/docs/#files-view-versions-of-a-file}
}
\seealso{
\code{\link[=box_dl(), box_read]{box_dl(), box_read()}}
}
