% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_file_versions.R
\name{box_previous_versions}
\alias{box_previous_versions}
\title{Get version information}
\usage{
box_previous_versions(file_id)
}
\arguments{
\item{file_id}{\code{numeric} or \code{character}, file ID at Box.}
}
\value{
\code{data.frame} describing previous versions of file.
}
\description{
\lifecycle{superseded} Superseded by \code{\link[=box_version_history]{box_version_history()}}.
}
