% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_read.R
\name{box_read}
\alias{box_read}
\alias{box_read_csv}
\alias{box_read_tsv}
\alias{box_read_json}
\alias{box_read_excel}
\title{Read an R object from a Box file}
\usage{
box_read(file_id, type = NULL, version_id = NULL, version_no = NULL,
  read_fun = rio::import, fread = FALSE, ...)

box_read_csv(file_id, ...)

box_read_tsv(file_id, ...)

box_read_json(file_id, ...)

box_read_excel(file_id, ...)
}
\arguments{
\item{file_id}{\code{numeric} or \code{character}, file ID at Box}

\item{type}{\code{character},
\href{http://en.wikipedia.org/wiki/Internet_media_type}{MIME type}
used to override the content type returned by the server.}

\item{version_id}{\code{character} or \code{numeric}, the \code{version_id} of the file}

\item{version_no}{\code{numeric}, version of the file you'd like to download
(starting at 1)}

\item{read_fun}{\code{function}, used to read (parse) the content into R; default
function is \code{\link[rio:import]{rio::import()}}}

\item{fread}{\code{logical}, indicates to use function \code{\link[data.table:fread]{data.table::fread()}}
to read CSV files}

\item{...}{additional arguments passed to \code{read_fun}}
}
\value{
The object returned by the function \code{read_fun}.
}
\description{
These functions are used to download a Box file, specified by \code{file_id}, then
attempt to parse its contents into memory as an R object. For
example, you may wish to read a Box CSV file as a \code{data.frame}.
}
\details{
This is a two-step process. The first is to download the contents
of the file, the second is to parse those contents into an R object.
The default parsing-function is \code{\link[rio:import]{rio::import()}}; the only modification of
its behavior is that a JSON file is not necessarily coerced to a
\code{data.frame}. In addition to \code{box_read()}, some specific helpers are
provided:

\describe{
\item{\code{box_read_csv()}}{parse a remote CSV file into a \code{data.frame} using \code{\link[=read.csv]{read.csv()}}}
\item{\code{box_read_tsv()}}{parse a remote TSV file into a \code{data.frame} using \code{\link[=read.delim]{read.delim()}}}
\item{\code{box_read_json()}}{parse a remote JSON file into a \code{list} using \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}}
\item{\code{box_read_excel()}}{parse a remote Microsoft Excel file into a \code{data.frame} using \code{\link[readxl:read_excel]{readxl::read_excel()}}}
}
}
\seealso{
\code{\link[=box_dl]{box_dl()}}, \code{\link[=box_save]{box_save()}}, \code{\link[=box_source]{box_source()}}
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
