% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_save_load.R
\name{box_save}
\alias{box_save}
\alias{box_load}
\alias{box_save_image}
\title{Download/upload an R workspace from/to a Box file}
\usage{
box_save(..., dir_id = box_getwd(), filename = ".RData",
  description = NULL, file_name)

box_save_image(dir_id = box_getwd(), filename = ".RData",
  description = NULL, file_name)

box_load(file_id)
}
\arguments{
\item{...}{objects to be saved, quoted or unquoted; passed to \code{\link[=save]{save()}}.}

\item{dir_id}{\code{numeric} or \code{character},
folder ID at Box}

\item{filename}{\code{character}, if supplied, an alternate filename
for the local version of the file.}

\item{description}{\code{character}, description caption for the file}

\item{file_name}{\code{character}, \strong{deprecated}: use \code{filename} instead}

\item{file_id}{\code{numeric} or \code{character}, file ID at Box}
}
\value{
\describe{
\item{\code{box_save()}}{S3 object with class \link[=boxr_S3_classes]{boxr_file_reference}}
\item{\code{box_save_image()}}{S3 object with class \link[=boxr_S3_classes]{boxr_file_reference}}
\item{\code{box_load()}}{From \code{\link[=load]{load()}}, a character vector of the names of objects
created, invisibly.}
}
}
\description{
Similar to \code{\link[=save]{save()}}, \code{\link[=save.image]{save.image()}}, and \code{\link[=load]{load()}}; these functions operate on
files at Box instead of on local files.
}
\details{
\describe{
\item{\code{box_save()}}{save object(s) using \code{\link[=save]{save()}}, write to Box file}
\item{\code{box_save_image()}}{save image using \code{\link[=save.image]{save.image()}}, write to Box file}
\item{\code{box_load()}}{read from Box file, load using \code{\link[=load]{load()}}}
}
}
\seealso{
\code{\link[=save]{save()}}, \code{\link[=save.image]{save.image()}}, \code{\link[=load]{load()}}
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
