% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_search.R
\name{box_search}
\alias{box_search}
\alias{box_search_files}
\alias{box_search_folders}
\alias{box_search_trash}
\title{Search the files in a box.com account}
\usage{
box_search(query, content_types = c("name", "description", "file_content",
  "comments", "tags"), type = c(NULL, "file", "folder", "weblink"),
  file_extensions = NULL, ancestor_folder_ids = NULL,
  created_at_range = NULL, updated_at_range = NULL, size_range = NULL,
  trash = FALSE, owner_user_ids = NULL, auto_paginate = TRUE, max = 200)

box_search_files(query, ...)

box_search_folders(query, ...)

box_search_trash(query, ...)
}
\arguments{
\item{query}{The search term that you'd like to use}

\item{content_types}{The content types you'd like to search. More than one 
can be supplied with a vector}

\item{type}{The type of object you'd like returned. The default of 
\code{NULL} return all possible types.}

\item{file_extensions}{A vector of strings containing the file extensions 
(without dots) by which to narrow your search.}

\item{ancestor_folder_ids}{If supplied, results are limited to one or more 
parent folders}

\item{created_at_range}{Optional. Limit search to a range of created at 
times. A vector of two dates (coercible via \code{\link{as.POSIXct}}).}

\item{updated_at_range}{Optional. Limit search to a range of updated at 
times. A vector of two dates (coercible via \code{\link{as.POSIXct}}).}

\item{size_range}{Optional. Limit search to a range of file sizes. A vector 
of two whole numbers (coerible via (coercible via \code{\link{as.numeric}} 
and \code{\link{round}})).}

\item{trash}{Should the search be limited to the trash folder? 
\code{\link{logical}}.}

\item{owner_user_ids}{Optional. Limit search to a files owned by a set of 
users. A vector if IDs.}

\item{auto_paginate}{\code{logical}. By default, the box.com will return a 
fixed number of search results per request. By setting auto_paginate to 
\code{TRUE}, boxr will keep making new requests untill all search results 
are in.}

\item{max}{\code{numeric}. Upper limit on the number of search results
returned (protective measure for users with large numbers of files).}

\item{...}{Parameters passed to \code{box_search}}
}
\description{
Search the files in a box.com account
}

