% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{box_setwd}
\alias{box_setwd}
\alias{box_getwd}
\title{Get/set Box default working-directory}
\usage{
box_setwd(dir_id)

box_getwd()
}
\arguments{
\item{dir_id}{\code{numeric} or \code{character}, folder ID at Box.}
}
\value{
\describe{
\item{\code{box_getwd()}}{\code{numeric}, ID for working folder at Box.}
\item{\code{box_setwd()}}{\code{invisible(NULL)}, called for side-effects.}
}
}
\description{
Similar to \code{\link[=getwd]{getwd()}} and \code{\link[=setwd]{setwd()}},
these functions get and set the folder ID of the working directory
at \href{https://www.box.com}{box.com}.

This folder ID is also stored in \code{\link[=boxr_options]{boxr_options()}}.
}
\seealso{
\code{\link[=box_ls]{box_ls()}} to list files in a Box directory,
\code{\link[=box_fetch]{box_fetch()}}/\code{\link[=box_push]{box_push()}} to download/upload directories from/to Box
}
