% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{box_setwd}
\alias{box_getwd}
\alias{box_setwd}
\title{Get/Set Default box.com directory/folder}
\usage{
box_setwd(dir_id)

box_getwd()
}
\arguments{
\item{dir_id}{The id of the folder you'd like to set as your default}
}
\value{
\code{box_getwd} returns the id of the default folder.
  \code{box_setwd} does nothing and is ussed for its side-effects.
}
\description{
Providing analgous functionality for the base \code{\bold{R}}
  functions \code{\link{getwd}} and \code{\link{setwd}}, these functions set
  and retrieve a default box.com dir_id, stored in
  \code{\link{boxr_options}}.
}
\seealso{
\code{\link{box_ls}} to examine the contents of a remote directory,
  and \code{\link{box_fetch}}/\code{\link{box_push}} for synchorizing them.
}

