% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr_source.R
\name{box_source}
\alias{box_source}
\title{Execute R code stored on box.com}
\usage{
box_source(file_id)
}
\arguments{
\item{file_id}{The box.com id for the file that you'd like to download}
}
\description{
\code{box_source} will download a remote R script stored on box.com, and then
attempt to execute it, using \code{\link{source}}.
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\seealso{
\code{\link{box_dl}} for saving files to disk,
  \code{\link{box_save}} for working with R workspaces, and
  \code{\link{box_read}} for reading files into memory as R objects.
}

