% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_write.R
\name{box_write}
\alias{box_write}
\title{Write R Objects to Files Remotely Hosted on box.com}
\usage{
box_write(x, filename, dir_id = box_getwd(), write_fun = rio::export, ...)
}
\arguments{
\item{x}{An R object}

\item{filename}{The name for the file to be uploaded}

\item{dir_id}{If uploading, the box.com folder id that you'd like to upload
to.}

\item{write_fun}{The function used to write the R object to a file}

\item{...}{Additional arguments passed to \code{read_fun}}
}
\value{
A refernce to to the file uploaded
}
\description{
{
  file format. \code{read_fun} is used to convert R objects to files, which
  by default is the \code{\link{export}} function from the \code{rio} 
  package.
  
  Note: \code{box_write} is for writing files in standard formats to box.com.
  To upload R objects as \code{.RData} files, see \code{\link{box_save}}.
}
}

