% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr-package.R
\docType{package}
\name{boxr-package}
\alias{boxr}
\alias{boxr-package}
\title{boxr: access the Box API}
\description{
A lightweight, \emph{opinionated}, high-level R interface for the \href{https://developer.box.com/docs}{box.com} API.
}
\details{
This package has a documentation-website (created using \href{https://pkgdown.r-lib.org/}{pkgdown}), containing:
\itemize{
\item \href{https://r-box.github.io/boxr/}{README}.
\item \href{https://r-box.github.io/boxr/articles/boxr.html}{Get-started article},
also accessible from R: \code{vignette("boxr")}.
\item \href{https://r-box.github.io/boxr/reference/index.html}{Function reference}.
}

The boxr source-repository is at GitHub: \url{https://github.com/r-box/boxr}.

If you find anything you think might be a bug, please report it as a
\href{https://github.com/r-box/boxr/issues}{GitHub issue}!

Happy hacking! :)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-box.github.io/boxr/}
  \item \url{https://github.com/r-box/boxr}
  \item Report bugs at \url{https://github.com/r-box/boxr/issues}
}

}
\author{
\strong{Maintainer}: Ian Lyttle \email{ijlyttle@me.com} (\href{https://orcid.org/0000-0001-9962-4849}{ORCID})

Authors:
\itemize{
  \item Brendan Rocks \email{foss@brendanrocks.com}
  \item Nathan Day \email{nathancday@gmail.com} (\href{https://orcid.org/0000-0002-6714-8611}{ORCID})
}

Other contributors:
\itemize{
  \item Vincent Fulco [contributor]
  \item Alec Wong [contributor]
  \item Alex Brodersen [contributor]
}

}
\keyword{internal}
