% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr__internal_package_homepage.R
\docType{package}
\name{boxr-package}
\alias{boxr}
\alias{boxr-package}
\title{boxr}
\description{
A light-weight, high-level R interface for the box.com API
}
\details{
An introduction to the package, and a short getting started guide (including 
how to authorize boxr to work with your box.com account), can be found in the
package's vignette. Simply run

\code{vignette("boxr")}

and it should open in your browser.

To see a list of the functions which \code{boxr} provides, click on the 
"Index" hyperlink at the bottom of this page.

\code{boxr}'s home on the internet is currently GitHub: 
\url{https://github.com/brendan-r/boxr}. If you find anything you think might
be a bug, do report it as a 
\href{https://github.com/brendan-R/boxr/issues}{GitHub issue}!

Happy hacking! :)
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\keyword{package}

