% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{boxr_options}
\alias{boxr_options}
\title{Display global options for boxr}
\usage{
boxr_options()
}
\value{
A \code{list} of the options available
}
\description{
Run \code{boxr_options()} to see what it's possible to set globally.
}
\details{
Options can be set in the usual way using \code{\link{options}}, and include:

\describe{
  \item{\code{box.verbose}}{
    Should boxr print to the console using \code{\link{cat}}? This is
    slightly 'rude' package behaviour, and may cause problems if using the
    \code{knitr} package.
  }
  \item{\code{box.wd}}{
    A list containg the name and id of the default box.com directory
  }
  \item{\code{box.token}}{
    The token object used for authentication
  }
}
}
\seealso{
\code{\link{box_setwd}} for another way to set the default box.com
  directory
}

