% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxr_misc.R
\name{boxr_options}
\alias{boxr_options}
\title{Display global options for boxr}
\usage{
boxr_options()
}
\value{
A \code{list} of the options available
}
\description{
Run \code{boxr_options()} to see what it's possible to set globally.
}
\details{
You should really write out all the options here, and explain them!
}

