% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxr__internal_diff_and_delete.R,
%   R/boxr__internal_diff_and_uldl.R
\name{deleteRemoteObjects}
\alias{deleteRemoteObjects}
\alias{downloadDirFiles}
\alias{uploadDirFiles}
\title{Single Directory Operations for Downloading and Uploading multiple Files}
\usage{
deleteRemoteObjects(dir_id, local_dir = getwd())

downloadDirFiles(dir_id, local_dir = getwd(), overwrite = TRUE,
  dir_str = getwd())

uploadDirFiles(dir_id, local_dir = getwd(), overwrite = TRUE)
}
\arguments{
\item{dir_id}{The box.com id for the folder that you'd like to query}

\item{local_dir}{The local directory which you'd like \code{dir_id} to
correspond to. If you're not interested in mapping local to hosted
directories, this isn't terribly important.}
}
\value{
\code{TRUE} for a successful sync, \code{NULL} if the box.com folder
is empty.
}
\description{
Download or upload the contents of a box.com directory, not including
subdirectories
}
\keyword{internal}
